refactor: BikeServiceDirectoryFetcher config
Refactor the configuration loading to support dependency injection and
add missing documentation.
