+++ src/main/java/org/opentripplanner/standalone/config/sandbox/BikeRentalServiceDirectoryFetcherConfig.java
@@ -1,6 +1,7 @@
-package org.opentripplanner.standalone.config.ext.bikerentalservicedirectoryfetcher;
+package org.opentripplanner.standalone.config.sandbox;
 
 import org.opentripplanner.standalone.config.NodeAdapter;
+import org.opentripplanner.ext.bikerentalservicedirectory.api.BikeRentalServiceDirectoryFetcherParameters;
 
 public class BikeRentalServiceDirectoryFetcherConfig {
   public static BikeRentalServiceDirectoryFetcherParameters create(NodeAdapter c) {
